backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pandas
numpy
python-dp
sqlalchemy
aiosqlite
psycopg2-binary
boto3
google-cloud-storage
matplotlib
diffprivlib
httpx
python-dotenv